requirements.txt
pylint
pytest
black
requestsp